Create PS-UploadToSharePoint.ps1
PowerShell script from Microsoft "Upload on-premises content to SharePoint Online using PowerShell cmdlets"
First version without any modifications except for correction of 2 minor mistakes in article and added notation about what steps in the article are used. Most likely there will be a version 2 soon with some added features and modifications

Be aware that you must run this on a domain joined computer with a user that have access to the files. Target are OneDrive site for a user in my example, but it will work against any site where user have permission.

Go to my blog for a video of the scripts being run https://olavtvedt.blogspot.com

Link to original:
https://support.office.com/en-us/article/upload-on-premises-content-to-sharepoint-online-using-powershell-cmdlets-555049c6-15ef-45a6-9a1f-a1ef673b867c?