Created Test-MigrationReadyness.ps1
This is an add-on to make Test-OneDrivePath.ps1 only check against only specified folders (and its subfolders) instead of all folders in the specified path. I have used the SharePoint Migration Tool's input file as the source for folders since this makes me able to combine one input file with the migration itself. But any CSV file with the first column being a path can be used.

All scanned top folders (Users home folder) which contain errors will result in a log file. Top folders with no errors will not get a result file. The process itself is logged in the Test-MigrationReadyness.log. Here you will find info about when the process started and ended for each top folder.