Create Set-AzureAppOwner
The owner of an app (App registration and Enterprise App) might be empty as a result of the original creator/owner no longer exist. Or you just need to add an additional user/group as owner.

This script to set owner on a list of Apps in Azure, based on the csv input file.